Replace the read cache with pure request coalescing
The read cache stored response bodies for 60s (10min for the catalog) and
replayed them. Its correctness rested on a cache key covering every dimension
a response depends on, an open-ended invariant that has already been broken
twice: a path-only key served page 1 of the plugin catalog to page-2 requests
(codex looped on its own cursor: 3,598 requests, 15 GB RSS, kernel panics) and
served one account's installed plugins to another. The loop guard, TTL,
eviction, and entry cap all existed to fence that stored state.

Delete the stored state and keep only singleflight: identical concurrent
requests to the plugin polling endpoints share one upstream fetch, nothing is
retained afterward, so there is no staleness window and nothing to guard.
Non-concurrent polls pass through, which is traffic parity with no proxy at
all. The full-request flightKey and its key-completeness property test carry
over unchanged.

Catalog aggregation stays (codex OOMs by pages walked and its disk cache is
cold for any proxy base URL), but a walk that fails midway or hits the
page/byte cap now returns an error instead of the pages collected so far. A
partial merged catalog carries no continuation token, so clients cache it as
complete; codex pinned 173- and 209-entry catalogs (of 2,285) for 3 hours
after upstream throttled a walk. A 502 gets retried; wrong data does not.

Net -271 lines.
